refactoring | #102 | add abstract class methods
