+++ packages/cli-ui/src/components/KillPort/style.module.scss
@@ -2,7 +2,7 @@
 	background: #fff;
 	border-radius: 6px;
 	box-shadow: 0 4px 15px rgba(0,0,0,.05);
-	width: 400px;
+	min-width: 400px;
 	height: 200px;
 	display: inline-block;
